refactor(saga): move saga modules and tests into package workspaces
