+++ packages/saga-runtime/package.json
@@ -16,7 +16,11 @@
   "scripts": {
     "build": "tsc -p tsconfig.json",
     "typecheck": "tsc -p tsconfig.json --noEmit",
-    "test": "bun test",
+    "test": "bun test ./test",
     "lint": "bunx tsc -p tsconfig.json --noEmit"
+  },
+  "dependencies": {
+    "@redemeine/aggregate": "workspace:*",
+    "@redemeine/saga": "workspace:*"
   }
 }
