+++ packages/saga-runtime/src/index.ts
@@ -1,2 +1,7 @@
-// @redemeine/saga-runtime package entrypoint placeholder.
-export {};
+declare const require: (id: string) => any;
+
+const sagaPackage = require('@redemeine/saga');
+
+export const createSagaDispatchContext = sagaPackage.createSagaDispatchContext as any;
+export const runSagaHandler = sagaPackage.runSagaHandler as any;
+export * from './createSagaAggregate';
