+++ src/redemeine.ts
@@ -11,4 +11,3 @@ export * from './Depot';
 export * from './types';
 export * from './Contract';
 export * from './projections';
-export * from './sagas';
